Fix modern WGSL syntax
The webgpu examples have ancient WGSL annotation syntax. This patch cleans that up, and also renames the fragment shader so the name doesn't collide.

Note that it's presenting a black screen, I didn't dig into why we're not seeing correct rendering. But the motivation for this is to capture performance traces in Xcode, so that's good enough.